Hadoop, MapReduce, Python
Find the most commonly uttered keywords in all of the Presidential
speeches made from the birth of the nation till now.
Also determine, from the current speech, compared to the previous few
years, if the President's plan for current year are different in
general.
Used Hadoop HDFS to store the addresses, Python for web scraping all the
addresses and MapReduce to process the huge amount of data.